Split validation module
