+++ weasyprint/css/validation/properties.py
@@ -1,0 +1,1751 @@
+"""
+    weasyprint.css.validation.properties
+    ------------------------------------
+
+    Validate properties.
+    See http://www.w3.org/TR/CSS21/propidx.html and various CSS3 modules.
+
+    :copyright: Copyright 2011-2018 Simon Sapin and contributors, see AUTHORS.
+    :license: BSD, see LICENSE for details.
+
+"""
+
+
+import math
+from urllib.parse import unquote
+
+from tinycss2.color3 import parse_color
+
+from .. import computed_values
+from ...formatting_structure import counters
+from ...images import LinearGradient, RadialGradient
+from ...logger import LOGGER
+from ..properties import KNOWN_PROPERTIES, Dimension
+from .functions import parse_function
+from .utils import (
+    InvalidValues, comma_separated_list, get_angle,
+    get_keyword, get_length, get_resolution, get_single_keyword,
+    remove_whitespace, safe_urljoin, single_keyword, single_token,
+    split_on_comma)
+
+
+PREFIX = '-weasy-'
+PROPRIETARY = set()
+UNSTABLE = set()
+
+ZERO_PERCENT = Dimension(0, '%')
+FIFTY_PERCENT = Dimension(50, '%')
+HUNDRED_PERCENT = Dimension(100, '%')
+BACKGROUND_POSITION_PERCENTAGES = {
+    'top': ZERO_PERCENT,
+    'left': ZERO_PERCENT,
+    'center': FIFTY_PERCENT,
+    'bottom': HUNDRED_PERCENT,
+    'right': HUNDRED_PERCENT,
+}
+
+CONTENT_QUOTE_KEYWORDS = {
+    'open-quote': (True, True),
+    'close-quote': (False, True),
+    'no-open-quote': (True, False),
+    'no-close-quote': (False, False),
+}
+DIRECTION_KEYWORDS = {
+    # ('angle', radians)  0 upwards, then clockwise
+    ('to', 'top'): ('angle', 0),
+    ('to', 'right'): ('angle', math.pi / 2),
+    ('to', 'bottom'): ('angle', math.pi),
+    ('to', 'left'): ('angle', math.pi * 3 / 2),
+    # ('corner', keyword)
+    ('to', 'top', 'left'): ('corner', 'top_left'),
+    ('to', 'left', 'top'): ('corner', 'top_left'),
+    ('to', 'top', 'right'): ('corner', 'top_right'),
+    ('to', 'right', 'top'): ('corner', 'top_right'),
+    ('to', 'bottom', 'left'): ('corner', 'bottom_left'),
+    ('to', 'left', 'bottom'): ('corner', 'bottom_left'),
+    ('to', 'bottom', 'right'): ('corner', 'bottom_right'),
+    ('to', 'right', 'bottom'): ('corner', 'bottom_right'),
+}
+
+
+# Yes/no validators for non-shorthand properties
+# Maps property names to functions taking a property name and a value list,
+# returning a value or None for invalid.
+# For properties that take a single value, that value is returned by itself
+# instead of a list.
+# TODO: fix this: Also transform values: keyword and URLs are returned as
+# strings.
+PROPERTIES = {}
+
+
+# Validators
+
+def property(property_name=None, proprietary=False, unstable=False,
+             wants_base_url=False):
+    """Decorator adding a function to the ``PROPERTIES``.
+
+    The name of the property covered by the decorated function is set to
+    ``property_name`` if given, or is inferred from the function name
+    (replacing underscores by hyphens).
+
+    :param proprietary:
+        Proprietary (vendor-specific, non-standard) are prefixed: anchors can
+        for example be set using ``-weasy-anchor: attr(id)``.
+        See https://www.w3.org/TR/CSS/#proprietary
+    :param unstable:
+        Mark properties that are defined in specifications that didn't reach
+        the Candidate Recommandation stage. They can be used both
+        vendor-prefixed or unprefixed.
+        See https://www.w3.org/TR/CSS/#unstable-syntax
+    :param wants_base_url:
+        The function takes the stylesheet’s base URL as an additional
+        parameter.
+
+    """
+    def decorator(function):
+        """Add ``function`` to the ``PROPERTIES``."""
+        if property_name is None:
+            name = function.__name__.replace('_', '-')
+        else:
+            name = property_name
+        assert name in KNOWN_PROPERTIES, name
+        assert name not in PROPERTIES, name
+
+        function.wants_base_url = wants_base_url
+        PROPERTIES[name] = function
+        print('lol', name)
+        if proprietary:
+            PROPRIETARY.add(name)
+        if unstable:
+            UNSTABLE.add(name)
+        return function
+    return decorator
+
+
+# TODO: fix this function
+def validate_content_list_token(base_url, token, for_content_box):
+    """Validation for a single token of <content-list> used in GCPM.
+
+    Not really.
+    GCPM <content-list> =
+        [ <string> | contents | <image> | <quote> | <target> | <leader()> ]+
+    (Draft, 24 January 2018. Really a DRAFT. Not an RFC. Not a SPEC.
+    BTW: The current Draft GCPM ``string-set`` value =
+        none | [ <custom-ident> <string>+ ]#
+
+    So. This is the validation for tokens that make sense in
+    css properties ``string-set``, ``bookmark-label`` and  ``content``:
+
+    <modified-content-list> = [
+      <string> | attr() | <counter> | <target> |
+      <content> |
+      url() | <quote> | string() | leader()
+    ]+
+
+    :param for_content_box: controls which tokens are valid
+
+    Valid tokens when ``for_content_box`` ==
+
+    - True (called from/for css property 'content':
+
+      <string> | attr() | <counter> | <target> |
+      url() | <quote> | string() | leader()
+
+      The final decision whether a token is valid is the job of
+      computed_values.content()
+
+    - False (called from/for css properties 'string-set', 'bookmark-label':
+      <string> | attr() | <counter> | <target> |
+      <content>
+
+    Return (type, content) or False for invalid tokens.
+
+    """
+
+    def validate_target_token(token):
+        """ validate first parameter of ``target-*()``-token
+            returns ['attr', '<attrname>' ]
+                 or ['STRING', '<anchorname>'] when valid
+            evaluation of the anchorname is job of compute()
+        """
+        # TODO: what about ``attr(href url)`` ?
+        if isinstance(token, str):
+            # url() or "string" given
+            # verify #anchor is done in compute()
+            # if token.value.startswith('#'):
+            return ['STRING', token]
+        function = parse_function(token)
+        if function:
+            name, args = function
+            params = [a.type for a in args]
+            values = [getattr(a, 'value', a) for a in args]
+            if name == 'attr' and params == ['ident']:
+                return [name, values[0]]
+
+    if for_content_box:
+        quote_type = CONTENT_QUOTE_KEYWORDS.get(get_keyword(token))
+        if quote_type is not None:
+            return ('QUOTE', quote_type)
+    else:
+        if get_keyword(token) == 'contents':
+            return ('content', 'text')
+    type_ = token.type
+    if type_ == 'string':
+        return ('STRING', token.value)
+    if for_content_box:
+        if type_ == 'url':
+            return ('URI', safe_urljoin(base_url, token.value))
+    function = parse_function(token)
+    if not function:
+        # to pass unit test `test_boxes.test_before_after`
+        # the log string must contain "invalid value"
+        raise InvalidValues('invalid value/unsupported token ´%s\´' % (token,))
+
+    name, args = function
+    # known functions in 'content', 'string-set' and 'bookmark-label':
+    valid_functions = ['attr',
+                       'counter', 'counters',
+                       'target-counter', 'target-counters', 'target-text']
+    # 'content'
+    if for_content_box:
+        valid_functions += ['string',
+                            'leader']
+    else:
+        valid_functions += ['content']
+    unsupported_functions = ['leader']
+    if name not in valid_functions:
+        # to pass unit test `test_boxes.test_before_after`
+        # the log string must contain "invalid value"
+        raise InvalidValues('invalid value: function `%s()`' % (name))
+    if name in unsupported_functions:
+        # suppress -- not (yet) implemented, no error
+        LOGGER.warn('\'%s()\' not (yet) supported', name)
+        return ('STRING', '')
+
+    prototype = (name, [a.type for a in args])
+    args = [getattr(a, 'value', a) for a in args]
+    if prototype == ('attr', ['ident']):
+        # TODO: what about ``attr(href url)`` ?
+        return (name, args[0])
+    elif prototype in (('content', []), ('content', ['ident', ])):
+        if not args:
+            return (name, 'text')
+        elif args[0] in ('text', 'after', 'before', 'first-letter'):
+            return (name, args[0])
+    elif prototype in (('counter', ['ident']),
+                       ('counters', ['ident', 'string'])):
+        args.append('decimal')
+        return (name, args)
+    elif prototype in (('counter', ['ident', 'ident']),
+                       ('counters', ['ident', 'string', 'ident'])):
+        style = args[-1]
+        if style in ('none', 'decimal') or style in counters.STYLES:
+            return (name, args)
+    elif prototype in (('string', ['ident']),
+                       ('string', ['ident', 'ident'])):
+        if len(args) > 1:
+            args[1] = args[1].lower()
+            if args[1] not in ('first', 'start', 'last', 'first-except'):
+                raise InvalidValues()
+        return (name, args)
+    # target-counter() = target-counter(
+    #    [ <string> | <url> ] , <custom-ident> ,
+    #    <counter-style>? )
+    elif name == 'target-counter':
+        if prototype in ((name, ['url', 'ident']),
+                         (name, ['url', 'ident', 'ident']),
+                         (name, ['string', 'ident']),
+                         (name, ['string', 'ident', 'ident']),
+                         (name, ['function', 'ident']),
+                         (name, ['function', 'ident', 'ident'])):
+            # default style
+            if len(args) == 2:
+                args.append('decimal')
+            # accept "#anchorname" and attr(x)
+            retval = validate_target_token(args.pop(0))
+            if retval is None:
+                raise InvalidValues()
+            style = args[-1]
+            if style in ('none', 'decimal') or style in counters.STYLES:
+                return (name, retval + args)
+    # target-counters() = target-counters(
+    #    [ <string> | <url> ] , <custom-ident> , <string> ,
+    #    <counter-style>? )
+    elif name == 'target-counters':
+        if prototype in ((name, ['url', 'ident', 'string']),
+                         (name, ['url', 'ident', 'string', 'ident']),
+                         (name, ['string', 'ident', 'string']),
+                         (name, ['string', 'ident', 'string', 'ident']),
+                         (name, ['function', 'ident', 'string']),
+                         (name, ['function', 'ident', 'string', 'ident'])):
+            # default style
+            if len(args) == 3:
+                args.append('decimal')
+            # accept "#anchorname" and attr(x)
+            retval = validate_target_token(args.pop(0))
+            if retval is None:
+                raise InvalidValues()
+            style = args[-1]
+            if style in ('none', 'decimal') or style in counters.STYLES:
+                return (name, retval + args)
+    # target-text() = target-text(
+    #    [ <string> | <url> ] ,
+    #    [ content | before | after | first-letter ]? )
+    elif name == 'target-text':
+        if prototype in ((name, ['url']),
+                         (name, ['url', 'ident']),
+                         (name, ['string']),
+                         (name, ['string', 'ident']),
+                         (name, ['function']),
+                         (name, ['function', 'ident'])):
+            if len(args) == 1:
+                args.append('content')
+            # accept "#anchorname" and attr(x)
+            retval = validate_target_token(args.pop(0))
+            if retval is None:
+                raise InvalidValues()
+            style = args[-1]
+            # hint: the syntax isn't stable yet!
+            if style in ('content', 'after', 'before', 'first-letter'):
+                # build.TEXT_CONTENT_EXTRACTORS needs 'text'
+                # TODO: should we define
+                # TEXT_CONTENT_EXTRACTORS['content'] == box_text ?
+                if style == 'content':
+                    args[-1] = 'text'
+                return (name, retval + args)
+
+
+def validate_non_shorthand(base_url, name, tokens, required=False):
+    """Default validator for non-shorthand properties."""
+    if not required and name not in KNOWN_PROPERTIES:
+        hyphens_name = name.replace('_', '-')
+        if hyphens_name in KNOWN_PROPERTIES:
+            raise InvalidValues('did you mean %s?' % hyphens_name)
+        else:
+            raise InvalidValues('unknown property')
+
+    if not required and name not in PROPERTIES:
+        raise InvalidValues('property not supported yet')
+
+    keyword = get_single_keyword(tokens)
+    if keyword in ('initial', 'inherit'):
+        value = keyword
+    else:
+        function = PROPERTIES[name]
+        if function.wants_base_url:
+            value = function(tokens, base_url)
+        else:
+            value = function(tokens)
+        if value is None:
+            raise InvalidValues
+    return ((name, value),)
+
+
+@property()
+@comma_separated_list
+@single_keyword
+def background_attachment(keyword):
+    """``background-attachment`` property validation."""
+    return keyword in ('scroll', 'fixed', 'local')
+
+
+@property('background-color')
+@property('border-top-color')
+@property('border-right-color')
+@property('border-bottom-color')
+@property('border-left-color')
+@property('column-rule-color')
+@single_token
+def other_colors(token):
+    return parse_color(token)
+
+
+@property()
+@single_token
+def outline_color(token):
+    if get_keyword(token) == 'invert':
+        return 'currentColor'
+    else:
+        return parse_color(token)
+
+
+@property()
+@single_keyword
+def border_collapse(keyword):
+    return keyword in ('separate', 'collapse')
+
+
+@property()
+@single_keyword
+def empty_cells(keyword):
+    """``empty-cells`` property validation."""
+    return keyword in ('show', 'hide')
+
+
+@property('color')
+@single_token
+def color(token):
+    """``*-color`` and ``color`` properties validation."""
+    result = parse_color(token)
+    if result == 'currentColor':
+        return 'inherit'
+    else:
+        return result
+
+
+@property('background-image', wants_base_url=True)
+@comma_separated_list
+@single_token
+def background_image(token, base_url):
+    if token.type != 'function':
+        return image_url([token], base_url)
+    arguments = split_on_comma(remove_whitespace(token.arguments))
+    name = token.lower_name
+    if name in ('linear-gradient', 'repeating-linear-gradient'):
+        direction, color_stops = parse_linear_gradient_parameters(arguments)
+        if color_stops:
+            return 'linear-gradient', LinearGradient(
+                [parse_color_stop(stop) for stop in color_stops],
+                direction, 'repeating' in name)
+    elif name in ('radial-gradient', 'repeating-radial-gradient'):
+        result = parse_radial_gradient_parameters(arguments)
+        if result is not None:
+            shape, size, position, color_stops = result
+        else:
+            shape = 'ellipse'
+            size = 'keyword', 'farthest-corner'
+            position = 'left', FIFTY_PERCENT, 'top', FIFTY_PERCENT
+            color_stops = arguments
+        if color_stops:
+            return 'radial-gradient', RadialGradient(
+                [parse_color_stop(stop) for stop in color_stops],
+                shape, size, position, 'repeating' in name)
+
+
+def parse_linear_gradient_parameters(arguments):
+    first_arg = arguments[0]
+    if len(first_arg) == 1:
+        angle = get_angle(first_arg[0])
+        if angle is not None:
+            return ('angle', angle), arguments[1:]
+    else:
+        result = DIRECTION_KEYWORDS.get(tuple(map(get_keyword, first_arg)))
+        if result is not None:
+            return result, arguments[1:]
+    return ('angle', math.pi), arguments  # Default direction is 'to bottom'
+
+
+def parse_radial_gradient_parameters(arguments):
+    shape = None
+    position = None
+    size = None
+    size_shape = None
+    stack = arguments[0][::-1]
+    while stack:
+        token = stack.pop()
+        keyword = get_keyword(token)
+        if keyword == 'at':
+            position = background_position.single_value(stack[::-1])
+            if position is None:
+                return
+            break
+        elif keyword in ('circle', 'ellipse') and shape is None:
+            shape = keyword
+        elif keyword in ('closest-corner', 'farthest-corner',
+                         'closest-side', 'farthest-side') and size is None:
+            size = 'keyword', keyword
+        else:
+            if stack and size is None:
+                length_1 = get_length(token, percentage=True)
+                length_2 = get_length(stack[-1], percentage=True)
+                if None not in (length_1, length_2):
+                    size = 'explicit', (length_1, length_2)
+                    size_shape = 'ellipse'
+                    stack.pop()
+            if size is None:
+                length_1 = get_length(token)
+                if length_1 is not None:
+                    size = 'explicit', (length_1, length_1)
+                    size_shape = 'circle'
+            if size is None:
+                return
+    if (shape, size_shape) in (('circle', 'ellipse'), ('circle', 'ellipse')):
+        return
+    return (
+        shape or size_shape or 'ellipse',
+        size or ('keyword', 'farthest-corner'),
+        position or ('left', FIFTY_PERCENT, 'top', FIFTY_PERCENT),
+        arguments[1:])
+
+
+def parse_color_stop(tokens):
+    if len(tokens) == 1:
+        color = parse_color(tokens[0])
+        if color is not None:
+            return color, None
+    elif len(tokens) == 2:
+        color = parse_color(tokens[0])
+        position = get_length(tokens[1], negative=True, percentage=True)
+        if color is not None and position is not None:
+            return color, position
+    raise InvalidValues
+
+
+@property('list-style-image', wants_base_url=True)
+@single_token
+def image_url(token, base_url):
+    """``*-image`` properties validation."""
+    if get_keyword(token) == 'none':
+        return 'none', None
+    if token.type == 'url':
+        return 'url', safe_urljoin(base_url, token.value)
+
+
+class CenterKeywordFakeToken(object):
+    type = 'ident'
+    lower_value = 'center'
+    unit = None
+
+
+@property(unstable=True)
+def transform_origin(tokens):
+    # TODO: parse (and ignore) a third value for Z.
+    return simple_2d_position(tokens)
+
+
+@property()
+@comma_separated_list
+def background_position(tokens):
+    """``background-position`` property validation.
+
+    See http://dev.w3.org/csswg/css3-background/#the-background-position
+
+    """
+    result = simple_2d_position(tokens)
+    if result is not None:
+        pos_x, pos_y = result
+        return 'left', pos_x, 'top', pos_y
+
+    if len(tokens) == 4:
+        keyword_1 = get_keyword(tokens[0])
+        keyword_2 = get_keyword(tokens[2])
+        length_1 = get_length(tokens[1], percentage=True)
+        length_2 = get_length(tokens[3], percentage=True)
+        if length_1 and length_2:
+            if (keyword_1 in ('left', 'right') and
+                    keyword_2 in ('top', 'bottom')):
+                return keyword_1, length_1, keyword_2, length_2
+            if (keyword_2 in ('left', 'right') and
+                    keyword_1 in ('top', 'bottom')):
+                return keyword_2, length_2, keyword_1, length_1
+
+    if len(tokens) == 3:
+        length = get_length(tokens[2], percentage=True)
+        if length is not None:
+            keyword = get_keyword(tokens[1])
+            other_keyword = get_keyword(tokens[0])
+        else:
+            length = get_length(tokens[1], percentage=True)
+            other_keyword = get_keyword(tokens[2])
+            keyword = get_keyword(tokens[0])
+
+        if length is not None:
+            if other_keyword == 'center':
+                if keyword in ('top', 'bottom'):
+                    return 'left', FIFTY_PERCENT, keyword, length
+                if keyword in ('left', 'right'):
+                    return keyword, length, 'top', FIFTY_PERCENT
+            elif (keyword in ('left', 'right') and
+                    other_keyword in ('top', 'bottom')):
+                return keyword, length, other_keyword, ZERO_PERCENT
+            elif (keyword in ('top', 'bottom') and
+                    other_keyword in ('left', 'right')):
+                return other_keyword, ZERO_PERCENT, keyword, length
+
+
+def simple_2d_position(tokens):
+    """Common syntax of background-position and transform-origin."""
+    if len(tokens) == 1:
+        tokens = [tokens[0], CenterKeywordFakeToken]
+    elif len(tokens) != 2:
+        return None
+
+    token_1, token_2 = tokens
+    length_1 = get_length(token_1, percentage=True)
+    length_2 = get_length(token_2, percentage=True)
+    if length_1 and length_2:
+        return length_1, length_2
+    keyword_1, keyword_2 = map(get_keyword, tokens)
+    if length_1 and keyword_2 in ('top', 'center', 'bottom'):
+        return length_1, BACKGROUND_POSITION_PERCENTAGES[keyword_2]
+    elif length_2 and keyword_1 in ('left', 'center', 'right'):
+            return BACKGROUND_POSITION_PERCENTAGES[keyword_1], length_2
+    elif (keyword_1 in ('left', 'center', 'right') and
+          keyword_2 in ('top', 'center', 'bottom')):
+        return (BACKGROUND_POSITION_PERCENTAGES[keyword_1],
+                BACKGROUND_POSITION_PERCENTAGES[keyword_2])
+    elif (keyword_1 in ('top', 'center', 'bottom') and
+          keyword_2 in ('left', 'center', 'right')):
+        # Swap tokens. They need to be in (horizontal, vertical) order.
+        return (BACKGROUND_POSITION_PERCENTAGES[keyword_2],
+                BACKGROUND_POSITION_PERCENTAGES[keyword_1])
+
+
+@property()
+@comma_separated_list
+def background_repeat(tokens):
+    """``background-repeat`` property validation."""
+    keywords = tuple(map(get_keyword, tokens))
+    if keywords == ('repeat-x',):
+        return ('repeat', 'no-repeat')
+    if keywords == ('repeat-y',):
+        return ('no-repeat', 'repeat')
+    if keywords in (('no-repeat',), ('repeat',), ('space',), ('round',)):
+        return keywords * 2
+    if len(keywords) == 2 and all(
+            k in ('no-repeat', 'repeat', 'space', 'round')
+            for k in keywords):
+        return keywords
+
+
+@property()
+@comma_separated_list
+def background_size(tokens):
+    """Validation for ``background-size``."""
+    if len(tokens) == 1:
+        token = tokens[0]
+        keyword = get_keyword(token)
+        if keyword in ('contain', 'cover'):
+            return keyword
+        if keyword == 'auto':
+            return ('auto', 'auto')
+        length = get_length(token, negative=False, percentage=True)
+        if length:
+            return (length, 'auto')
+    elif len(tokens) == 2:
+        values = []
+        for token in tokens:
+            length = get_length(token, negative=False, percentage=True)
+            if length:
+                values.append(length)
+            elif get_keyword(token) == 'auto':
+                values.append('auto')
+        if len(values) == 2:
+            return tuple(values)
+
+
+@property('background-clip')
+@property('background-origin')
+@comma_separated_list
+@single_keyword
+def box(keyword):
+    """Validation for the ``<box>`` type used in ``background-clip``
+    and ``background-origin``."""
+    return keyword in ('border-box', 'padding-box', 'content-box')
+
+
+@property()
+def border_spacing(tokens):
+    """Validator for the `border-spacing` property."""
+    lengths = [get_length(token, negative=False) for token in tokens]
+    if all(lengths):
+        if len(lengths) == 1:
+            return (lengths[0], lengths[0])
+        elif len(lengths) == 2:
+            return tuple(lengths)
+
+
+@property('border-top-right-radius')
+@property('border-bottom-right-radius')
+@property('border-bottom-left-radius')
+@property('border-top-left-radius')
+def border_corner_radius(tokens):
+    """Validator for the `border-*-radius` properties."""
+    lengths = [
+        get_length(token, negative=False, percentage=True) for token in tokens]
+    if all(lengths):
+        if len(lengths) == 1:
+            return (lengths[0], lengths[0])
+        elif len(lengths) == 2:
+            return tuple(lengths)
+
+
+@property('border-top-style')
+@property('border-right-style')
+@property('border-left-style')
+@property('border-bottom-style')
+@property('column-rule-style')
+@single_keyword
+def border_style(keyword):
+    """``border-*-style`` properties validation."""
+    return keyword in ('none', 'hidden', 'dotted', 'dashed', 'double',
+                       'inset', 'outset', 'groove', 'ridge', 'solid')
+
+
+@property('break-before')
+@property('break-after')
+@single_keyword
+def break_before_after(keyword):
+    """``break-before`` and ``break-after`` properties validation."""
+    # 'always' is defined as an alias to 'page' in multi-column
+    # https://www.w3.org/TR/css3-multicol/#column-breaks
+    return keyword in ('auto', 'avoid', 'avoid-page', 'page', 'left', 'right',
+                       'recto', 'verso', 'avoid-column', 'column', 'always')
+
+
+@property()
+@single_keyword
+def break_inside(keyword):
+    """``break-inside`` property validation."""
+    return keyword in ('auto', 'avoid', 'avoid-page', 'avoid-column')
+
+
+@property(unstable=True)
+@single_token
+def page(token):
+    """``page`` property validation."""
+    if token.type == 'ident':
+        return 'auto' if token.lower_value == 'auto' else token.value
+
+
+@property("bleed-left")
+@property("bleed-right")
+@property("bleed-top")
+@property("bleed-bottom")
+@single_token
+def bleed(token):
+    """``bleed`` property validation."""
+    keyword = get_keyword(token)
+    if keyword == 'auto':
+        return 'auto'
+    else:
+        return get_length(token)
+
+
+@property()
+def marks(tokens):
+    """``marks`` property validation."""
+    if len(tokens) == 2:
+        keywords = [get_keyword(token) for token in tokens]
+        if 'crop' in keywords and 'cross' in keywords:
+            return keywords
+    elif len(tokens) == 1:
+        keyword = get_keyword(tokens[0])
+        if keyword in ('crop', 'cross'):
+            return [keyword]
+        elif keyword == 'none':
+            return 'none'
+
+
+@property('outline-style')
+@single_keyword
+def outline_style(keyword):
+    """``outline-style`` properties validation."""
+    return keyword in ('none', 'dotted', 'dashed', 'double', 'inset',
+                       'outset', 'groove', 'ridge', 'solid')
+
+
+@property('border-top-width')
+@property('border-right-width')
+@property('border-left-width')
+@property('border-bottom-width')
+@property('column-rule-width')
+@property('outline-width')
+@single_token
+def border_width(token):
+    """Border, column rule and outline widths properties validation."""
+    length = get_length(token, negative=False)
+    if length:
+        return length
+    keyword = get_keyword(token)
+    if keyword in ('thin', 'medium', 'thick'):
+        return keyword
+
+
+@property()
+@single_token
+def column_width(token):
+    """``column-width`` property validation."""
+    length = get_length(token, negative=False)
+    if length:
+        return length
+    keyword = get_keyword(token)
+    if keyword == 'auto':
+        return keyword
+
+
+@property()
+@single_keyword
+def column_span(keyword):
+    """``column-span`` property validation."""
+    # TODO: uncomment this when it is supported
+    # return keyword in ('all', 'none')
+
+
+@property()
+@single_keyword
+def box_sizing(keyword):
+    """Validation for the ``box-sizing`` property from css3-ui"""
+    return keyword in ('padding-box', 'border-box', 'content-box')
+
+
+@property()
+@single_keyword
+def caption_side(keyword):
+    """``caption-side`` properties validation."""
+    return keyword in ('top', 'bottom')
+
+
+@property()
+@single_keyword
+def clear(keyword):
+    """``clear`` property validation."""
+    return keyword in ('left', 'right', 'both', 'none')
+
+
+@property()
+@single_token
+def clip(token):
+    """Validation for the ``clip`` property."""
+    function = parse_function(token)
+    if function:
+        name, args = function
+        if name == 'rect' and len(args) == 4:
+            values = []
+            for arg in args:
+                if get_keyword(arg) == 'auto':
+                    values.append('auto')
+                else:
+                    length = get_length(arg)
+                    if length:
+                        values.append(length)
+            if len(values) == 4:
+                return tuple(values)
+    if get_keyword(token) == 'auto':
+        return ()
+
+
+@property(wants_base_url=True)
+def content(tokens, base_url):
+    """``content`` property validation.
+    TODO: should become a @comma_separated_list to validate
+          CSS3 <content-replacement>
+    """
+    keyword = get_single_keyword(tokens)
+    if keyword in ('normal', 'none'):
+        return keyword
+    parsed_tokens = [validate_content_token(base_url, v) for v in tokens]
+    if None not in parsed_tokens:
+        return parsed_tokens
+
+
+def validate_content_token(base_url, token):
+    """Validation for a single token for the ``content`` property.
+
+    Return (type, content) or False for invalid tokens.
+
+    """
+    return validate_content_list_token(
+        base_url, token, for_content_box=True)
+
+
+@property()
+def counter_increment(tokens):
+    """``counter-increment`` property validation."""
+    return counter(tokens, default_integer=1)
+
+
+@property()
+def counter_reset(tokens):
+    """``counter-reset`` property validation."""
+    return counter(tokens, default_integer=0)
+
+
+def counter(tokens, default_integer):
+    """``counter-increment`` and ``counter-reset`` properties validation."""
+    if get_single_keyword(tokens) == 'none':
+        return ()
+    tokens = iter(tokens)
+    token = next(tokens, None)
+    assert token, 'got an empty token list'
+    results = []
+    while token is not None:
+        if token.type != 'ident':
+            return  # expected a keyword here
+        counter_name = token.value
+        if counter_name in ('none', 'initial', 'inherit'):
+            raise InvalidValues('Invalid counter name: ' + counter_name)
+        token = next(tokens, None)
+        if token is not None and (
+                token.type == 'number' and token.int_value is not None):
+            # Found an integer. Use it and get the next token
+            integer = token.int_value
+            token = next(tokens, None)
+        else:
+            # Not an integer. Might be the next counter name.
+            # Keep `token` for the next loop iteration.
+            integer = default_integer
+        results.append((counter_name, integer))
+    return tuple(results)
+
+
+@property('top')
+@property('right')
+@property('left')
+@property('bottom')
+@property('margin-top')
+@property('margin-right')
+@property('margin-bottom')
+@property('margin-left')
+@single_token
+def lenght_precentage_or_auto(token):
+    """``margin-*`` properties validation."""
+    length = get_length(token, percentage=True)
+    if length:
+        return length
+    if get_keyword(token) == 'auto':
+        return 'auto'
+
+
+@property('height')
+@property('width')
+@single_token
+def width_height(token):
+    """Validation for the ``width`` and ``height`` properties."""
+    length = get_length(token, negative=False, percentage=True)
+    if length:
+        return length
+    if get_keyword(token) == 'auto':
+        return 'auto'
+
+
+@property()
+@single_token
+def column_gap(token):
+    """Validation for the ``column-gap`` property."""
+    length = get_length(token, negative=False)
+    if length:
+        return length
+    keyword = get_keyword(token)
+    if keyword == 'normal':
+        return keyword
+
+
+@property()
+@single_keyword
+def column_fill(keyword):
+    """``column-fill`` property validation."""
+    return keyword in ('auto', 'balance')
+
+
+@property()
+@single_keyword
+def direction(keyword):
+    """``direction`` property validation."""
+    return keyword in ('ltr', 'rtl')
+
+
+@property()
+@single_keyword
+def display(keyword):
+    """``display`` property validation."""
+    return keyword in (
+        'inline', 'block', 'inline-block', 'list-item', 'none',
+        'table', 'inline-table', 'table-caption',
+        'table-row-group', 'table-header-group', 'table-footer-group',
+        'table-row', 'table-column-group', 'table-column', 'table-cell',
+        'flex', 'inline-flex')
+
+
+@property('float')
+@single_keyword
+def float_(keyword):  # XXX do not hide the "float" builtin
+    """``float`` property validation."""
+    return keyword in ('left', 'right', 'none')
+
+
+@property()
+@comma_separated_list
+def font_family(tokens):
+    """``font-family`` property validation."""
+    if len(tokens) == 1 and tokens[0].type == 'string':
+        return tokens[0].value
+    elif tokens and all(token.type == 'ident' for token in tokens):
+        return ' '.join(token.value for token in tokens)
+
+
+@property()
+@single_keyword
+def font_kerning(keyword):
+    return keyword in ('auto', 'normal', 'none')
+
+
+@property()
+@single_token
+def font_language_override(token):
+    keyword = get_keyword(token)
+    if keyword == 'normal':
+        return keyword
+    elif token.type == 'string':
+        return token.value
+
+
+@property()
+def font_variant_ligatures(tokens):
+    if len(tokens) == 1:
+        keyword = get_keyword(tokens[0])
+        if keyword in ('normal', 'none'):
+            return keyword
+    values = []
+    couples = (
+        ('common-ligatures', 'no-common-ligatures'),
+        ('historical-ligatures', 'no-historical-ligatures'),
+        ('discretionary-ligatures', 'no-discretionary-ligatures'),
+        ('contextual', 'no-contextual'))
+    all_values = []
+    for couple in couples:
+        all_values.extend(couple)
+    for token in tokens:
+        if token.type != 'ident':
+            return None
+        if token.value in all_values:
+            concurrent_values = [
+                couple for couple in couples if token.value in couple][0]
+            if any(value in values for value in concurrent_values):
+                return None
+            else:
+                values.append(token.value)
+        else:
+            return None
+    if values:
+        return tuple(values)
+
+
+@property()
+@single_keyword
+def font_variant_position(keyword):
+    return keyword in ('normal', 'sub', 'super')
+
+
+@property()
+@single_keyword
+def font_variant_caps(keyword):
+    return keyword in (
+        'normal', 'small-caps', 'all-small-caps', 'petite-caps',
+        'all-petite-caps', 'unicase', 'titling-caps')
+
+
+@property()
+def font_variant_numeric(tokens):
+    if len(tokens) == 1:
+        keyword = get_keyword(tokens[0])
+        if keyword == 'normal':
+            return keyword
+    values = []
+    couples = (
+        ('lining-nums', 'oldstyle-nums'),
+        ('proportional-nums', 'tabular-nums'),
+        ('diagonal-fractions', 'stacked-fractions'),
+        ('ordinal',), ('slashed-zero',))
+    all_values = []
+    for couple in couples:
+        all_values.extend(couple)
+    for token in tokens:
+        if token.type != 'ident':
+            return None
+        if token.value in all_values:
+            concurrent_values = [
+                couple for couple in couples if token.value in couple][0]
+            if any(value in values for value in concurrent_values):
+                return None
+            else:
+                values.append(token.value)
+        else:
+            return None
+    if values:
+        return tuple(values)
+
+
+@property()
+def font_feature_settings(tokens):
+    """``font-feature-settings`` property validation."""
+    if len(tokens) == 1 and get_keyword(tokens[0]) == 'normal':
+        return 'normal'
+
+    @comma_separated_list
+    def font_feature_settings_list(tokens):
+        feature, value = None, None
+
+        if len(tokens) == 2:
+            tokens, token = tokens[:-1], tokens[-1]
+            if token.type == 'ident':
+                value = {'on': 1, 'off': 0}.get(token.value)
+            elif (token.type == 'number' and
+                    token.int_value is not None and token.int_value >= 0):
+                value = token.int_value
+        elif len(tokens) == 1:
+            value = 1
+
+        if len(tokens) == 1:
+            token, = tokens
+            if token.type == 'string' and len(token.value) == 4:
+                if all(0x20 <= ord(letter) <= 0x7f for letter in token.value):
+                    feature = token.value
+
+        if feature is not None and value is not None:
+            return feature, value
+
+    return font_feature_settings_list(tokens)
+
+
+@property()
+@single_keyword
+def font_variant_alternates(keyword):
+    # TODO: support other values
+    # See https://www.w3.org/TR/css-fonts-3/#font-variant-caps-prop
+    return keyword in ('normal', 'historical-forms')
+
+
+@property()
+def font_variant_east_asian(tokens):
+    if len(tokens) == 1:
+        keyword = get_keyword(tokens[0])
+        if keyword == 'normal':
+            return keyword
+    values = []
+    couples = (
+        ('jis78', 'jis83', 'jis90', 'jis04', 'simplified', 'traditional'),
+        ('full-width', 'proportional-width'),
+        ('ruby',))
+    all_values = []
+    for couple in couples:
+        all_values.extend(couple)
+    for token in tokens:
+        if token.type != 'ident':
+            return None
+        if token.value in all_values:
+            concurrent_values = [
+                couple for couple in couples if token.value in couple][0]
+            if any(value in values for value in concurrent_values):
+                return None
+            else:
+                values.append(token.value)
+        else:
+            return None
+    if values:
+        return tuple(values)
+
+
+@property()
+@single_token
+def font_size(token):
+    """``font-size`` property validation."""
+    length = get_length(token, negative=False, percentage=True)
+    if length:
+        return length
+    font_size_keyword = get_keyword(token)
+    if font_size_keyword in ('smaller', 'larger'):
+        raise InvalidValues('value not supported yet')
+    if font_size_keyword in computed_values.FONT_SIZE_KEYWORDS:
+        # or keyword in ('smaller', 'larger')
+        return font_size_keyword
+
+
+@property()
+@single_keyword
+def font_style(keyword):
+    """``font-style`` property validation."""
+    return keyword in ('normal', 'italic', 'oblique')
+
+
+@property()
+@single_keyword
+def font_stretch(keyword):
+    """Validation for the ``font-stretch`` property."""
+    return keyword in (
+        'ultra-condensed', 'extra-condensed', 'condensed', 'semi-condensed',
+        'normal',
+        'semi-expanded', 'expanded', 'extra-expanded', 'ultra-expanded')
+
+
+@property()
+@single_token
+def font_weight(token):
+    """``font-weight`` property validation."""
+    keyword = get_keyword(token)
+    if keyword in ('normal', 'bold', 'bolder', 'lighter'):
+        return keyword
+    if token.type == 'number' and token.int_value is not None:
+        if token.int_value in (100, 200, 300, 400, 500, 600, 700, 800, 900):
+            return token.int_value
+
+
+@property(unstable=True)
+@single_token
+def image_resolution(token):
+    # TODO: support 'snap' and 'from-image'
+    return get_resolution(token)
+
+
+@property('letter-spacing')
+@property('word-spacing')
+@single_token
+def spacing(token):
+    """Validation for ``letter-spacing`` and ``word-spacing``."""
+    if get_keyword(token) == 'normal':
+        return 'normal'
+    length = get_length(token)
+    if length:
+        return length
+
+
+@property()
+@single_token
+def line_height(token):
+    """``line-height`` property validation."""
+    if get_keyword(token) == 'normal':
+        return 'normal'
+    if token.type == 'number' and token.value >= 0:
+        return Dimension(token.value, None)
+    if token.type == 'percentage' and token.value >= 0:
+        return Dimension(token.value, '%')
+    elif token.type == 'dimension' and token.value >= 0:
+        return get_length(token)
+
+
+@property()
+@single_keyword
+def list_style_position(keyword):
+    """``list-style-position`` property validation."""
+    return keyword in ('inside', 'outside')
+
+
+@property()
+@single_keyword
+def list_style_type(keyword):
+    """``list-style-type`` property validation."""
+    return keyword in ('none', 'decimal') or keyword in counters.STYLES
+
+
+@property('min-width')
+@property('min-height')
+@single_token
+def min_width_height(token):
+    """``min-width`` and ``min-height`` properties validation."""
+    # See https://www.w3.org/TR/css-flexbox-1/#min-size-auto
+    keyword = get_keyword(token)
+    if keyword == 'auto':
+        return keyword
+    else:
+        return length_or_precentage([token])
+
+
+@property('padding-top')
+@property('padding-right')
+@property('padding-bottom')
+@property('padding-left')
+@single_token
+def length_or_precentage(token):
+    """``padding-*`` properties validation."""
+    length = get_length(token, negative=False, percentage=True)
+    if length:
+        return length
+
+
+@property('max-width')
+@property('max-height')
+@single_token
+def max_width_height(token):
+    """Validation for max-width and max-height"""
+    length = get_length(token, negative=False, percentage=True)
+    if length:
+        return length
+    if get_keyword(token) == 'none':
+        return Dimension(float('inf'), 'px')
+
+
+@property()
+@single_token
+def opacity(token):
+    """Validation for the ``opacity`` property."""
+    if token.type == 'number':
+        return min(1, max(0, token.value))
+
+
+@property()
+@single_token
+def z_index(token):
+    """Validation for the ``z-index`` property."""
+    if get_keyword(token) == 'auto':
+        return 'auto'
+    if token.type == 'number' and token.int_value is not None:
+        return token.int_value
+
+
+@property('orphans')
+@property('widows')
+@single_token
+def orphans_widows(token):
+    """Validation for the ``orphans`` and ``widows`` properties."""
+    if token.type == 'number' and token.int_value is not None:
+        value = token.int_value
+        if value >= 1:
+            return value
+
+
+@property()
+@single_token
+def column_count(token):
+    """Validation for the ``column-count`` property."""
+    if token.type == 'number' and token.int_value is not None:
+        value = token.int_value
+        if value >= 1:
+            return value
+    if get_keyword(token) == 'auto':
+        return 'auto'
+
+
+@property()
+@single_keyword
+def overflow(keyword):
+    """Validation for the ``overflow`` property."""
+    return keyword in ('auto', 'visible', 'hidden', 'scroll')
+
+
+@property()
+@single_keyword
+def position(keyword):
+    """``position`` property validation."""
+    return keyword in ('static', 'relative', 'absolute', 'fixed')
+
+
+@property()
+def quotes(tokens):
+    """``quotes`` property validation."""
+    if (tokens and len(tokens) % 2 == 0 and
+            all(v.type == 'string' for v in tokens)):
+        strings = tuple(token.value for token in tokens)
+        # Separate open and close quotes.
+        # eg.  ('«', '»', '“', '”')  -> (('«', '“'), ('»', '”'))
+        return strings[::2], strings[1::2]
+
+
+@property()
+@single_keyword
+def table_layout(keyword):
+    """Validation for the ``table-layout`` property"""
+    if keyword in ('fixed', 'auto'):
+        return keyword
+
+
+@property()
+@single_keyword
+def text_align(keyword):
+    """``text-align`` property validation."""
+    return keyword in ('left', 'right', 'center', 'justify')
+
+
+@property()
+def text_decoration(tokens):
+    """``text-decoration`` property validation."""
+    keywords = [get_keyword(v) for v in tokens]
+    if keywords == ['none']:
+        return 'none'
+    if all(keyword in ('underline', 'overline', 'line-through', 'blink')
+            for keyword in keywords):
+        unique = set(keywords)
+        if len(unique) == len(keywords):
+            # No duplicate
+            # blink is accepted but ignored
+            # "Conforming user agents may simply not blink the text."
+            return frozenset(unique - set(['blink']))
+
+
+@property()
+@single_token
+def text_indent(token):
+    """``text-indent`` property validation."""
+    length = get_length(token, percentage=True)
+    if length:
+        return length
+
+
+@property()
+@single_keyword
+def text_transform(keyword):
+    """``text-align`` property validation."""
+    return keyword in (
+        'none', 'uppercase', 'lowercase', 'capitalize', 'full-width')
+
+
+@property()
+@single_token
+def vertical_align(token):
+    """Validation for the ``vertical-align`` property"""
+    length = get_length(token, percentage=True)
+    if length:
+        return length
+    keyword = get_keyword(token)
+    if keyword in ('baseline', 'middle', 'sub', 'super',
+                   'text-top', 'text-bottom', 'top', 'bottom'):
+        return keyword
+
+
+@property()
+@single_keyword
+def visibility(keyword):
+    """``white-space`` property validation."""
+    return keyword in ('visible', 'hidden', 'collapse')
+
+
+@property()
+@single_keyword
+def white_space(keyword):
+    """``white-space`` property validation."""
+    return keyword in ('normal', 'pre', 'nowrap', 'pre-wrap', 'pre-line')
+
+
+@property()
+@single_keyword
+def overflow_wrap(keyword):
+    """``overflow-wrap`` property validation."""
+    return keyword in ('normal', 'break-word')
+
+
+@property()
+@single_token
+def flex_basis(token):
+    """``flex-basis`` property validation."""
+    basis = width_height([token])
+    if basis is not None:
+        return basis
+    if get_keyword(token) == 'content':
+        return 'content'
+
+
+@property()
+@single_keyword
+def flex_direction(keyword):
+    """``flex-direction`` property validation."""
+    return keyword in ('row', 'row-reverse', 'column', 'column-reverse')
+
+
+@property('flex-grow')
+@property('flex-shrink')
+@single_token
+def flex_grow_shrink(token):
+    if token.type == 'number':
+        return token.value
+
+
+@property()
+@single_token
+def order(token):
+    if token.type == 'number' and token.int_value is not None:
+        return token.int_value
+
+
+@property()
+@single_keyword
+def flex_wrap(keyword):
+    """``flex-wrap`` property validation."""
+    return keyword in ('nowrap', 'wrap', 'wrap-reverse')
+
+
+@property()
+@single_keyword
+def justify_content(keyword):
+    """``justify-content`` property validation."""
+    return keyword in (
+        'flex-start', 'flex-end', 'center', 'space-between', 'space-around')
+
+
+@property()
+@single_keyword
+def align_items(keyword):
+    """``align-items`` property validation."""
+    return keyword in (
+        'flex-start', 'flex-end', 'center', 'baseline', 'stretch')
+
+
+@property()
+@single_keyword
+def align_self(keyword):
+    """``align-self`` property validation."""
+    return keyword in (
+        'auto', 'flex-start', 'flex-end', 'center', 'baseline', 'stretch')
+
+
+@property()
+@single_keyword
+def align_content(keyword):
+    """``align-content`` property validation."""
+    return keyword in (
+        'flex-start', 'flex-end', 'center', 'space-between', 'space-around',
+        'stretch')
+
+
+@property(unstable=True)
+@single_keyword
+def image_rendering(keyword):
+    """Validation for ``image-rendering``."""
+    return keyword in ('auto', 'crisp-edges', 'pixelated')
+
+
+@property(unstable=True)
+def size(tokens):
+    """``size`` property validation.
+
+    See http://www.w3.org/TR/css3-page/#page-size-prop
+
+    """
+    lengths = [get_length(token, negative=False) for token in tokens]
+    if all(lengths):
+        if len(lengths) == 1:
+            return (lengths[0], lengths[0])
+        elif len(lengths) == 2:
+            return tuple(lengths)
+
+    keywords = [get_keyword(v) for v in tokens]
+    if len(keywords) == 1:
+        keyword = keywords[0]
+        if keyword in computed_values.PAGE_SIZES:
+            return computed_values.PAGE_SIZES[keyword]
+        elif keyword in ('auto', 'portrait'):
+            return computed_values.INITIAL_PAGE_SIZE
+        elif keyword == 'landscape':
+            return computed_values.INITIAL_PAGE_SIZE[::-1]
+
+    if len(keywords) == 2:
+        if keywords[0] in ('portrait', 'landscape'):
+            orientation, page_size = keywords
+        elif keywords[1] in ('portrait', 'landscape'):
+            page_size, orientation = keywords
+        else:
+            page_size = None
+        if page_size in computed_values.PAGE_SIZES:
+            width_height = computed_values.PAGE_SIZES[page_size]
+            if orientation == 'portrait':
+                return width_height
+            else:
+                height, width = width_height
+                return width, height
+
+
+@property(proprietary=True)
+@single_token
+def anchor(token):
+    """Validation for ``anchor``."""
+    if get_keyword(token) == 'none':
+        return 'none'
+    function = parse_function(token)
+    if function:
+        name, args = function
+        prototype = (name, [a.type for a in args])
+        args = [getattr(a, 'value', a) for a in args]
+        if prototype == ('attr', ['ident']):
+            return (name, args[0])
+
+
+@property(proprietary=True, wants_base_url=True)
+@single_token
+def link(token, base_url):
+    """Validation for ``link``."""
+    if get_keyword(token) == 'none':
+        return 'none'
+    elif token.type == 'url':
+        if token.value.startswith('#'):
+            return 'internal', unquote(token.value[1:])
+        else:
+            return 'external', safe_urljoin(base_url, token.value)
+    function = parse_function(token)
+    if function:
+        name, args = function
+        prototype = (name, [a.type for a in args])
+        args = [getattr(a, 'value', a) for a in args]
+        if prototype == ('attr', ['ident']):
+            return (name, args[0])
+
+
+@property()
+@single_token
+def tab_size(token):
+    """Validation for ``tab-size``.
+
+    See https://www.w3.org/TR/css-text-3/#tab-size
+
+    """
+    if token.type == 'number' and token.int_value is not None:
+        value = token.int_value
+        if value >= 0:
+            return value
+    return get_length(token, negative=False)
+
+
+@property(unstable=True)
+@single_token
+def hyphens(token):
+    """Validation for ``hyphens``."""
+    keyword = get_keyword(token)
+    if keyword in ('none', 'manual', 'auto'):
+        return keyword
+
+
+@property(unstable=True)
+@single_token
+def hyphenate_character(token):
+    """Validation for ``hyphenate-character``."""
+    keyword = get_keyword(token)
+    if keyword == 'auto':
+        return '‐'
+    elif token.type == 'string':
+        return token.value
+
+
+@property(unstable=True)
+@single_token
+def hyphenate_limit_zone(token):
+    """Validation for ``hyphenate-limit-zone``."""
+    return get_length(token, negative=False, percentage=True)
+
+
+@property(unstable=True)
+def hyphenate_limit_chars(tokens):
+    """Validation for ``hyphenate-limit-chars``."""
+    if len(tokens) == 1:
+        token, = tokens
+        keyword = get_keyword(token)
+        if keyword == 'auto':
+            return (5, 2, 2)
+        elif token.type == 'number' and token.int_value is not None:
+            return (token.int_value, 2, 2)
+    elif len(tokens) == 2:
+        total, left = tokens
+        total_keyword = get_keyword(total)
+        left_keyword = get_keyword(left)
+        if total.type == 'number' and total.int_value is not None:
+            if left.type == 'number' and left.int_value is not None:
+                return (total.int_value, left.int_value, left.int_value)
+            elif left_keyword == 'auto':
+                return (total.value, 2, 2)
+        elif total_keyword == 'auto':
+            if left.type == 'number' and left.int_value is not None:
+                return (5, left.int_value, left.int_value)
+            elif left_keyword == 'auto':
+                return (5, 2, 2)
+    elif len(tokens) == 3:
+        total, left, right = tokens
+        if (
+            (get_keyword(total) == 'auto' or
+                (total.type == 'number' and total.int_value is not None)) and
+            (get_keyword(left) == 'auto' or
+                (left.type == 'number' and left.int_value is not None)) and
+            (get_keyword(right) == 'auto' or
+                (right.type == 'number' and right.int_value is not None))
+        ):
+            total = total.int_value if total.type == 'number' else 5
+            left = left.int_value if left.type == 'number' else 2
+            right = right.int_value if right.type == 'number' else 2
+            return (total, left, right)
+
+
+@property(proprietary=True)
+@single_token
+def lang(token):
+    """Validation for ``lang``."""
+    if get_keyword(token) == 'none':
+        return 'none'
+    function = parse_function(token)
+    if function:
+        name, args = function
+        prototype = (name, [a.type for a in args])
+        args = [getattr(a, 'value', a) for a in args]
+        if prototype == ('attr', ['ident']):
+            return (name, args[0])
+    elif token.type == 'string':
+        return ('string', token.value)
+
+
+@property(unstable=True, wants_base_url=True)
+def bookmark_label(tokens, base_url):
+    """Validation for ``bookmark-label``."""
+    parsed_tokens = tuple(validate_content_list_token(
+        base_url, v, for_content_box=False) for v in tokens)
+    if None not in parsed_tokens:
+        return parsed_tokens
+
+
+@property(unstable=True)
+@single_token
+def bookmark_level(token):
+    """Validation for ``bookmark-level``."""
+    if token.type == 'number' and token.int_value is not None:
+        value = token.int_value
+        if value >= 1:
+            return value
+    elif get_keyword(token) == 'none':
+        return 'none'
+
+
+@property(unstable=True, wants_base_url=True)
+@comma_separated_list
+def string_set(tokens, base_url):
+    """Validation for ``string-set``."""
+    if len(tokens) >= 2:
+        var_name = get_keyword(tokens[0])
+        parsed_tokens = tuple(
+            validate_content_list_token(
+                base_url, v, for_content_box=False) for v in tokens[1:])
+        if None not in parsed_tokens:
+            return (var_name, parsed_tokens)
+    elif tokens and tokens[0].value == 'none':
+        return 'none'
+
+
+@property(unstable=True)
+def transform(tokens):
+    if get_single_keyword(tokens) == 'none':
+        return ()
+    else:
+        transforms = []
+        for token in tokens:
+            function = parse_function(token)
+            if not function:
+                raise InvalidValues
+            name, args = function
+
+            if len(args) == 1:
+                angle = get_angle(args[0])
+                length = get_length(args[0], percentage=True)
+                if name in ('rotate', 'skewx', 'skewy') and angle:
+                    transforms.append((name, angle))
+                elif name in ('translatex', 'translate') and length:
+                    transforms.append((
+                        'translate', (length, computed_values.ZERO_PIXELS)))
+                elif name == 'translatey' and length:
+                    transforms.append((
+                        'translate', (computed_values.ZERO_PIXELS, length)))
+                elif name == 'scalex' and args[0].type == 'number':
+                    transforms.append(('scale', (args[0].value, 1)))
+                elif name == 'scaley' and args[0].type == 'number':
+                    transforms.append(('scale', (1, args[0].value)))
+                elif name == 'scale' and args[0].type == 'number':
+                    transforms.append(('scale', (args[0].value,) * 2))
+                else:
+                    raise InvalidValues
+            elif len(args) == 2:
+                if name == 'scale' and all(a.type == 'number' for a in args):
+                    transforms.append((name, tuple(arg.value for arg in args)))
+                else:
+                    lengths = tuple(
+                        get_length(token, percentage=True) for token in args)
+                    if name == 'translate' and all(lengths):
+                        transforms.append((name, lengths))
+                    else:
+                        raise InvalidValues
+            elif len(args) == 6 and name == 'matrix' and all(
+                    a.type == 'number' for a in args):
+                transforms.append((name, tuple(arg.value for arg in args)))
+            else:
+                raise InvalidValues
+        return tuple(transforms)
